+++ de.tonsias.basis.osgi.test/src/de/tonsias/basis/osgi/test/system/SingleValueServiceSystemTest.java
@@ -33,6 +33,7 @@ import de.tonsias.basis.model.interfaces.ISingleValue;
 import de.tonsias.basis.osgi.intf.IEventBrokerBridge.Type;
 import de.tonsias.basis.osgi.intf.IInstanzService;
 import de.tonsias.basis.osgi.intf.ISingleValueService;
+import de.tonsias.basis.osgi.intf.non.service.ChangeType;
 import de.tonsias.basis.osgi.intf.non.service.SingleValueEventConstants;
 import de.tonsias.basis.osgi.intf.non.service.SingleValueEventConstants.LinkedInstanzChangeEvent;
 import de.tonsias.basis.osgi.intf.non.service.SingleValueEventConstants.SingleValueDeleteEvent;
@@ -498,7 +499,7 @@ public class SingleValueServiceSystemTest {
 		LinkedInstanzChangeEvent data = _recorder.onlyDataOf(SingleValueEventConstants.INSTANZ_LIST_CHANGE,
 				LinkedInstanzChangeEvent.class);
 		assertThat(data._key(), is(value.getOwnKey()));
-		assertThat(data._changeType(), is(ValueEventConstants.ChangeType.ADD));
+		assertThat(data._changeType(), is(ChangeType.ADD));
 		assertThat(data._instanzKeys(), contains(secondOwner.getOwnKey()));
 
 		assertThat(value.getConnectedInstanzKeys(),